# target/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Parent pom providing dependency and plugin management for applications\n\t\tbuilt with Maven",
    "version" : "1.0.0",
    "title" : "forum-main"
  },
...
  "paths" : {
...
    "/api/crud/topic/topic/topics/{id}" : {
      "get" : {
        "tags" : [ "Topic" ],
        "summary" : "查看 Topic",
        "description" : "",
        "operationId" : "getTopic",
        "parameters" : [ {
          "name" : "",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Topic对象"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Topic" ],
        "summary" : "修改 Topic",
        "description" : "",
        "operationId" : "updateTopic",
        "parameters" : [ {
          "name" : "",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Topic对象"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Topic对象"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Topic" ],
        "summary" : "删除 Topic",
        "description" : "",
        "operationId" : "deleteTopic",
        "parameters" : [ {
          "name" : "",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Tip"
            }
          }
        }
      }
    },
...
    "/forum/comment/add" : {
      "post" : {
        "tags" : [ "forum" ],
        "summary" : "添加文章",
        "description" : "",
        "operationId" : "addComment",
        "parameters" : [ {
          "name" : "topicId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "content",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "topicUid",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseResult"
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "ResponseResult" : {
      "type" : "object",
      "properties" : {
        "success" : {
          "type" : "boolean"
        },
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "jumpUri" : {
          "type" : "string"
        }
      }
    },
...
    "Tip" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "Topic对象" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "话题id"
        },
        "getlId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "getlName" : {
          "type" : "string"
        },
        "uid" : {
          "type" : "integer",
          "format" : "int64"
        },
        "title" : {
          "type" : "string"
        },
        "time" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "更新时间"
        },
        "headUrl" : {
          "type" : "string"
        },
        "uname" : {
          "type" : "string"
        },
        "content" : {
          "type" : "string",
          "description" : "内容"
        },
        "summary" : {
          "type" : "string",
          "description" : "内容摘要"
        },
        "commentTotal" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "评论总数"
        },
        "istop" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "s是否置顶"
        },
        "audit" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "0为未审核，1为待人工审核，2为审核通过，3为不通过"
        },
        "contentFilter" : {
          "type" : "string"
        }
      }
    },
...
  }
}